Auditoria, deletar curso

+++ crud-Angular/src/app/course/containers/course/course.component.ts
@@ -10,6 +10,8 @@ import { ErrorDialogComponent } from '../../../shared/components/error-dialog/er
 import { SharedModule } from '../../../shared/shared.module';
 import { CourseListComponent } from '../../components/course-list/course-list.component';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { error } from 'node:console';
+import { ConfirmationDialogComponent } from '../../../shared/components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-course',
@@ -19,7 +21,7 @@ import { MatSnackBar } from '@angular/material/snack-bar';
   styleUrl: './course.component.scss',
 })
 export class CourseComponent implements OnInit {
-  course$: Observable<Course[]>;
+  course$: Observable<Course[]> | null = null;
 
   constructor(
     private courseService: CourseService,
@@ -28,6 +30,10 @@ export class CourseComponent implements OnInit {
     private routeActive: ActivatedRoute,
     public dialog: MatDialog
   ) {
+    this.refresh();
+  }
+
+  refresh() {
     this.course$ = this.courseService.list().pipe(
       catchError((error) => {
         this.onError('Erro ao carregar cursos!');
@@ -48,14 +54,31 @@ export class CourseComponent implements OnInit {
     this.router.navigate(['new'], { relativeTo: this.routeActive });
   }
 
-  onEdit(course: Course){
-    this.router.navigate(['edit', course._id], { relativeTo: this.routeActive });
+  onEdit(course: Course) {
+    this.router.navigate(['edit', course._id], {
+      relativeTo: this.routeActive,
+    });
   }
 
-  onDelete(course: Course){
-    this.courseService.delete(course._id).subscribe(
-      () =>
-        this.snackBar.open('Curso excluido com sucesso!', '',{duration: 5000} )
-    )
+  onDelete(course: Course) {
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: 'Tem certeza que deseja remover esse curso?',
+    });
+
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result) {
+        this.courseService.delete(course._id).subscribe({
+          next: () => {
+            this.refresh(),
+              this.snackBar.open('Curso excluido com sucesso!', '', {
+                duration: 5000,
+                verticalPosition: 'top',
+                horizontalPosition: 'center',
+              });
+          },
+          error: () => this.onError('Erro ao tentar remover curso!'),
+        });
+      }
+    });
   }
 }
